Implemented delete exercise for Workout and fixed constructor
Included a delete_exercise method that given an exercise name
could delete the exercise from the workout. Also fixed the multiple
constructors in the ruby file since I thought that ruby could do
that but actually cannot. The second constructor was the days_ago
constructor so in the future that will be saved by adding a question
during parsing on when the workout actually occured.