Fixed end issue, added modify_exercise method
There was an issue with containing an extra end in order for the code
to compile, however it would pass syntax check but none of the methods
worked. It turned out to be an issue in one of the previous methods
that propagated up to this point. Regardless, situation has been fixed.
The modify_exercise method seems to work based on preliminary testing
for modifying a specific WSR in a given exercise.