Completed serialization and deserialization of User, duck typing change
Conducted a basic test of serialization of User and looks to be working
properly, which is done using the Marshal package. The duck typing
change refers to simply modifying some of the code such that for checks
on whether the type is correct, it is done by duck typing (respond_to?)
rather than using dynamic typing (is_a?).