Complete template test, need to finish user tests
Template test was rather straightforward since there really aren't
any edge cases to it, because there is no functionality to remove
exercises within a template, only to delete the template entirely.
User test still needs to be done, and because there are a lot of cases
for it, it will be tested more completely than the other tests.