Include first code for session
The start session method is begining process of implementation,
starting with the add option for users. The add option then gets
separated into the each possible thing to add, and for each of those
there is a specific protocol. Therefore care needs to be taken on how
the function calls are structured.