Implement middle function for add_workout
When adding a workout, the user has the choice to either do a free
form version in which they list out every exercise, or they can use
a pre-defined template to save time. The add_workout method asks
the user which one they want. This continues to modularize the code
into distinct functions.