Complete user choice between template and freeform
The add_workout method now has the code that asks the user for whether
they want a template or to do it free form when adding a workout. The
templates were displayed as a number along with the template,
(ex. 1. Template0) and so there needed to be a check of if a user typed
in just the number 1, the code should handle and accept that as valid.