Add limited functionality to add workout template
In the process of completing the add_workout_with_template method.
The next thing to do is to implement the 'done' and 'more' options
when in the middle of adding the template workout. ask_for_wsr
was created as a new method because the add_workout methods were
getting too long since it's a lot of lines of code to ask for a weight,
and then sets, and then reps.