Finish implementation of adding a freeform workout
Initial tests of playing around with the code seems to be working fine.
Other than completing the freeform workout, which was in fact very
similar to the template version, I made changes on print statements and
user input so that users would put their input on the same line as the
question. This seemed better from a UX perspective.