Ensure that user cannot add same exercise twice
For a freeform workout, a user could potentially add the same
exercise name twice, which could lead to confusion for the workout
itself. The code now handles that edge case. If a user were to use
a template, adding a template also ensures that a template cannot have
the same exercise name twice.