Fix issue with no 'end-keyword'
In the add_exercise_for_workout method, there was an issue with the end
keyword as a syntax error. That has been fixed, as well as it looks like
the logic is correct on displaying workouts. But this is still from a
preliminary glance and it will be more rigorously tested once there are
more workouts available so as to print out more workouts.