Update workout to account for edge case
There's an edge case in workout such that if a user decided to add an
exercise with a name that is already in the workout, the workout should
not have two different exercises which are essentially the same. The
model now checks if the workout already contains the exercise and does
the necessary modifications.