Add feature to view workout
Moving onto the viewing option, the first quick one was to view a
workout. The next thing to do which won't take too long is to view a
template. Viewing an exercise across the different workouts will be a
bit more involved because the idea would be to add graphs to visually
see progress.