Add features for deleting workouts and templates
The deleting was rather straightforward since the code was exactly
the same as viewing, with the modification of doing a deletion instead
of printing out the string representation. It was decided that deleting
an exercise would be to tedious because it begs the question of what
deleting an exercise even means. That said, the next things to do are to
view an exercise graphically, and be able to export an exercise into
a spreadsheet.