Add functionality for different color text
colorize.rb is a String extension, taken from Stack overflow. It is
a useful feature to enhance the text and make it less dull. None
of the text has been modified yet but the functionality is there.
Next thing to still do is to add functionality for viewing exercises
and creating graphs from it