Add graphing feature for view exercise
Most of the graphing functionality for view_exercise is completed. One
thing that should be added for a better UI experience is to include
more labels other than just the first and the last one. There should
also be a check for if there's only one data point, in which case a
line graph doesn't really help.