Add coloring onto most test
The text helps to make some parts of the text stand out over others.
One part of the text that is not changed is lines that ask user for
input, which is fine because it still stands out by itself. Red is used
for errors, green is used for successful updates, and yellow is used
for any help text.